requirements.txt
django==4.2.2
Pillow==8.3.2
qrcode[pil]==7.0
